refactor: 친구 추가 관련 서비스 분리
